clean up executor tests and use tcp stream
