+++ crates/astria-conductor/src/executor/mod.rs
@@ -40,6 +40,9 @@ use crate::{
     types::SequencerBlockSubset,
 };
 
+#[cfg(test)]
+mod tests;
+
 /// The channel for sending commands to the executor task.
 pub(crate) type Sender = UnboundedSender<ExecutorCommand>;
 /// The channel the executor task uses to listen for commands.
@@ -98,13 +101,13 @@ pub(crate) struct Executor {
 
 impl Executor {
     pub(crate) async fn new(
-        client_url: &str,
+        server_addr: &str,
         chain_id: ChainId,
         disable_empty_block_execution: bool,
         cmd_rx: Receiver,
         shutdown: oneshot::Receiver<()>,
     ) -> Result<Self> {
-        let mut execution_rpc_client = ExecutionRpcClient::new(client_url)
+        let mut execution_rpc_client = ExecutionRpcClient::new(server_addr)
             .await
             .wrap_err("failed to create execution rpc client")?;
         let init_state_response = execution_rpc_client
@@ -302,309 +305,5 @@ impl Executor {
         self.sequencer_hash_to_execution_hash
             .remove(&sequencer_block_hash);
         Ok(())
-    }
-}
-
-#[cfg(test)]
-mod test {
-    use std::net::SocketAddr;
-
-    use color_eyre::eyre::Result;
-    use prost_types::Timestamp;
-    use proto::generated::execution::v1alpha1::{
-        execution_service_server::{
-            ExecutionService,
-            ExecutionServiceServer,
-        },
-        DoBlockRequest,
-        DoBlockResponse,
-        FinalizeBlockRequest,
-        FinalizeBlockResponse,
-        InitStateRequest,
-        InitStateResponse,
-    };
-    use sha2::Digest as _;
-    use tokio::{
-        sync::{
-            mpsc,
-            oneshot,
-        },
-        task::JoinHandle,
-    };
-    use tonic::transport::Server;
-
-    use super::*;
-
-    #[derive(Debug, Default)]
-    struct MockExecutionServer {}
-
-    #[async_trait::async_trait]
-    impl ExecutionClient for MockExecutionServer {
-        async fn call_init_state(&mut self) -> Result<InitStateResponse> {
-            unimplemented!("call_init_state")
-        }
-
-        async fn call_do_block(
-            &mut self,
-            _prev_block_hash: Vec<u8>,
-            _transactions: Vec<Vec<u8>>,
-            _timestamp: Option<Timestamp>,
-        ) -> Result<DoBlockResponse> {
-            unimplemented!("call_do_block")
-        }
-
-        async fn call_finalize_block(&mut self, _block_hash: Vec<u8>) -> Result<()> {
-            unimplemented!("call_finalize_block")
-        }
-    }
-
-    #[tonic::async_trait]
-    impl ExecutionService for MockExecutionServer {
-        async fn init_state(
-            &self,
-            _request: tonic::Request<InitStateRequest>,
-        ) -> std::result::Result<tonic::Response<InitStateResponse>, tonic::Status> {
-            let hasher = sha2::Sha256::new();
-            Ok(tonic::Response::new(InitStateResponse {
-                block_hash: hasher.finalize().to_vec(),
-            }))
-        }
-
-        async fn do_block(
-            &self,
-            _request: tonic::Request<DoBlockRequest>,
-        ) -> std::result::Result<tonic::Response<DoBlockResponse>, tonic::Status> {
-            let prev_block_hash = _request.into_inner().prev_block_hash;
-            let res = hash(&prev_block_hash);
-            Ok(tonic::Response::new(DoBlockResponse {
-                block_hash: res.to_vec(),
-            }))
-        }
-
-        async fn finalize_block(
-            &self,
-            _request: tonic::Request<FinalizeBlockRequest>,
-        ) -> std::result::Result<tonic::Response<FinalizeBlockResponse>, tonic::Status> {
-            Ok(tonic::Response::new(FinalizeBlockResponse {}))
-        }
-    }
-
-    pub(crate) struct MockExecution {
-        pub(crate) _server_handle: JoinHandle<()>,
-        pub(crate) addr: SocketAddr,
-    }
-
-    impl MockExecution {
-        pub(crate) async fn spawn() -> Self {
-            use tokio::net::TcpListener;
-            // randomly generating a local address
-            let listener = TcpListener::bind("127.0.0.1:0").await.unwrap();
-            let addr = listener.local_addr().unwrap();
-            drop(listener);
-
-            let server_handle = tokio::spawn(async move {
-                let _ = Server::builder()
-                    .add_service(ExecutionServiceServer::new(MockExecutionServer::default()))
-                    .serve(addr)
-                    .await;
-            });
-            tokio::time::sleep(tokio::time::Duration::from_millis(100)).await;
-            Self {
-                _server_handle: server_handle,
-                addr,
-            }
-        }
-
-        pub(crate) fn local_addr(&self) -> String {
-            format!("http://{}", self.addr)
-        }
-    }
-
-    fn hash(s: &[u8]) -> Vec<u8> {
-        let mut hasher = sha2::Sha256::new();
-        hasher.update(s);
-        hasher.finalize().to_vec()
-    }
-
-    fn get_test_block_subset() -> SequencerBlockSubset {
-        SequencerBlockSubset {
-            block_hash: hash(b"block1").try_into().unwrap(),
-            header: astria_sequencer_types::test_utils::default_header(),
-            rollup_transactions: vec![],
-        }
-    }
-
-    fn get_test_config() -> crate::Config {
-        crate::Config {
-            chain_id: "test".to_string(),
-            execution_rpc_url: "test".to_string(),
-            disable_finalization: false,
-            log: "test".to_string(),
-            disable_empty_block_execution: false,
-            celestia_node_url: "test".to_string(),
-            celestia_bearer_token: "test".to_string(),
-            sequencer_url: "test".to_string(),
-            initial_sequencer_block_height: 1,
-        }
-    }
-
-    #[tokio::test]
-    async fn execute_sequencer_block_without_txs() {
-        let mut cfg = get_test_config();
-        let chain_id = ChainId::new(cfg.chain_id.as_bytes().to_vec()).unwrap();
-
-        let execution_server = MockExecution::spawn().await;
-        cfg.execution_rpc_url = execution_server.local_addr();
-
-        let (_block_tx, block_rx) = mpsc::unbounded_channel();
-        let (_shutdown_tx, shutdown_rx) = oneshot::channel();
-        let mut executor = Executor::new(
-            &cfg.execution_rpc_url,
-            chain_id,
-            cfg.disable_empty_block_execution,
-            block_rx,
-            shutdown_rx,
-        )
-        .await
-        .unwrap();
-
-        let expected_exection_hash = hash(&executor.execution_state);
-        let block = get_test_block_subset();
-
-        let execution_block_hash = executor
-            .execute_block(block)
-            .await
-            .unwrap()
-            .expect("expected execution block hash");
-        assert_eq!(expected_exection_hash, execution_block_hash);
-    }
-
-    #[tokio::test]
-    async fn skip_sequencer_block_without_txs() {
-        let mut cfg = get_test_config();
-        let chain_id = ChainId::new(cfg.chain_id.as_bytes().to_vec()).unwrap();
-        cfg.disable_empty_block_execution = true;
-        let execution_server = MockExecution::spawn().await;
-        cfg.execution_rpc_url = execution_server.local_addr();
-
-        let (_block_tx, block_rx) = mpsc::unbounded_channel();
-        let (_shutdown_tx, shutdown_rx) = oneshot::channel();
-        let mut executor = Executor::new(
-            &cfg.execution_rpc_url,
-            chain_id,
-            cfg.disable_empty_block_execution,
-            block_rx,
-            shutdown_rx,
-        )
-        .await
-        .unwrap();
-
-        let block = get_test_block_subset();
-        let execution_block_hash = executor.execute_block(block).await.unwrap();
-        assert!(execution_block_hash.is_none());
-    }
-
-    #[tokio::test]
-    async fn execute_unexecuted_da_block_with_transactions() {
-        let mut cfg = get_test_config();
-        let chain_id = ChainId::new(cfg.chain_id.as_bytes().to_vec()).unwrap();
-
-        let execution_server = MockExecution::spawn().await;
-        cfg.execution_rpc_url = execution_server.local_addr();
-
-        let (_block_tx, block_rx) = mpsc::unbounded_channel();
-        let (_shutdown_tx, shutdown_rx) = oneshot::channel();
-        let mut executor = Executor::new(
-            &cfg.execution_rpc_url,
-            chain_id,
-            cfg.disable_empty_block_execution,
-            block_rx,
-            shutdown_rx,
-        )
-        .await
-        .unwrap();
-
-        let mut block = get_test_block_subset();
-        block.rollup_transactions.push(b"test_transaction".to_vec());
-
-        let expected_exection_hash = hash(&executor.execution_state);
-
-        executor
-            .handle_block_received_from_data_availability(block)
-            .await
-            .unwrap();
-
-        assert_eq!(expected_exection_hash, executor.execution_state);
-        // should be empty because 1 block was executed and finalized, which
-        // deletes it from the map
-        assert!(executor.sequencer_hash_to_execution_hash.is_empty());
-    }
-
-    #[tokio::test]
-    async fn skip_unexecuted_da_block_with_no_transactions() {
-        let mut cfg = get_test_config();
-        let chain_id = ChainId::new(cfg.chain_id.as_bytes().to_vec()).unwrap();
-        cfg.disable_empty_block_execution = true;
-
-        let execution_server = MockExecution::spawn().await;
-        cfg.execution_rpc_url = execution_server.local_addr();
-
-        let (_block_tx, block_rx) = mpsc::unbounded_channel();
-        let (_shutdown_tx, shutdown_rx) = oneshot::channel();
-        let mut executor = Executor::new(
-            &cfg.execution_rpc_url,
-            chain_id,
-            cfg.disable_empty_block_execution,
-            block_rx,
-            shutdown_rx,
-        )
-        .await
-        .unwrap();
-
-        let block: SequencerBlockSubset = get_test_block_subset();
-        let previous_execution_state = executor.execution_state.clone();
-
-        executor
-            .handle_block_received_from_data_availability(block)
-            .await
-            .unwrap();
-
-        assert_eq!(previous_execution_state, executor.execution_state,);
-        // should be empty because nothing was executed
-        assert!(executor.sequencer_hash_to_execution_hash.is_empty());
-    }
-
-    #[tokio::test]
-    async fn execute_unexecuted_da_block_with_no_transactions() {
-        let mut cfg = get_test_config();
-        let chain_id = ChainId::new(cfg.chain_id.as_bytes().to_vec()).unwrap();
-
-        let execution_server = MockExecution::spawn().await;
-        cfg.execution_rpc_url = execution_server.local_addr();
-
-        let (_block_tx, block_rx) = mpsc::unbounded_channel();
-        let (_shutdown_tx, shutdown_rx) = oneshot::channel();
-        let mut executor = Executor::new(
-            &cfg.execution_rpc_url,
-            chain_id,
-            cfg.disable_empty_block_execution,
-            block_rx,
-            shutdown_rx,
-        )
-        .await
-        .unwrap();
-
-        let block: SequencerBlockSubset = get_test_block_subset();
-        let expected_execution_state = hash(&executor.execution_state);
-
-        executor
-            .handle_block_received_from_data_availability(block)
-            .await
-            .unwrap();
-
-        assert_eq!(expected_execution_state, executor.execution_state);
-        // should be empty because block was executed and finalized, which
-        // deletes it from the map
-        assert!(executor.sequencer_hash_to_execution_hash.is_empty());
     }
 }
